Rewrite deparser with minimal implementation and add basic tests
